auth: updated endpoints

+++ src/main/java/com/rest/user_service/security/jwt/JwtUtils.java
@@ -1,7 +1,6 @@
 
 package com.rest.user_service.security.jwt;
 
-import com.rest.user_service.security.service.UserDetailsImpl;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
@@ -9,7 +8,6 @@ import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseCookie;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
 import org.springframework.web.util.WebUtils;
 
@@ -25,7 +23,7 @@ public class JwtUtils {
     private String jwtSecret;
 
     @Value("${spring.app.jwtExpirationMs}")
-    private Integer jwtExpirationMs;
+    private Long jwtExpirationMs;
 
     @Value("${spring.app.jwtCookie}")
     private String jwtCookie;
